rename: 컴포넌트, 페이지 폴더 이름 소문자로 변경
